[ETCM-1082] Move types used in storage to a single place
