+++ src/main/scala/io/iohk/ethereum/db/cache/AppCaches.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.db.cache
 
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.utils.Config
 
 trait AppCaches extends CacheComponent {
