+++ src/main/scala/io/iohk/ethereum/db/cache/CacheComponent.scala
@@ -1,11 +1,12 @@
 package io.iohk.ethereum.db.cache
 
-import io.iohk.ethereum.db.storage.NodeStorage
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 
 trait CacheComponent {
   val caches: Caches
 
   trait Caches {
-    val nodeCache: Cache[NodeStorage.NodeHash, NodeStorage.NodeEncoded]
+    val nodeCache: Cache[NodeHash, NodeEncoded]
   }
 }
