+++ src/main/scala/io/iohk/ethereum/db/components/Storages.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.db.components
 
 import io.iohk.ethereum.db.cache.AppCaches
 import io.iohk.ethereum.db.cache.LruCache
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.db.storage._
 import io.iohk.ethereum.db.storage.pruning.PruningMode
 import io.iohk.ethereum.utils.Config
