+++ src/main/scala/io/iohk/ethereum/db/storage/ArchiveNodeStorage.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.db.storage
 
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.mpt.NodesKeyValueStorage
 
 /** This class is used to store Nodes (defined in mpt/Node.scala), by using:
