+++ src/main/scala/io/iohk/ethereum/db/storage/MptStorage.scala
@@ -2,7 +2,7 @@ package io.iohk.ethereum.db.storage
 
 import akka.util.ByteString
 
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
 import io.iohk.ethereum.mpt.MerklePatriciaTrie.MissingRootNodeException
 import io.iohk.ethereum.mpt.MptNode
 import io.iohk.ethereum.mpt.MptTraversals
