+++ src/main/scala/io/iohk/ethereum/db/storage/NodeStorage.scala
@@ -8,8 +8,8 @@ import io.iohk.ethereum.db.cache.Cache
 import io.iohk.ethereum.db.dataSource.DataSource
 import io.iohk.ethereum.db.dataSource.DataSourceUpdateOptimized
 import io.iohk.ethereum.db.dataSource.RocksDbDataSource.IterationError
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 
 sealed trait NodesStorage extends {
   def get(key: NodeHash): Option[NodeEncoded]
@@ -71,9 +71,4 @@ class CachedNodeStorage(val storage: NodeStorage, val cache: Cache[NodeHash, Nod
   override type I = NodeStorage
   override def apply(cache: Cache[NodeHash, NodeEncoded], storage: NodeStorage): CachedNodeStorage =
     new CachedNodeStorage(storage, cache)
-}
-
-object NodeStorage {
-  type NodeHash = ByteString
-  type NodeEncoded = Array[Byte]
 }
