+++ src/main/scala/io/iohk/ethereum/db/storage/ReadOnlyNodeStorage.scala
@@ -2,8 +2,8 @@ package io.iohk.ethereum.db.storage
 
 import scala.collection.mutable
 
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.mpt.NodesKeyValueStorage
 
 /** This storage allows to read from another NodesKeyValueStorage but doesn't remove or upsert into database.
