+++ src/main/scala/io/iohk/ethereum/db/storage/StorageTypes.scala
@@ -1,0 +1,11 @@
+package io.iohk.ethereum.db.storage
+
+import akka.util.ByteString
+
+object StorageTypes {
+  type BlockHash = ByteString
+  type BlockHeaderHash = ByteString
+  type BlockBodyHash = ByteString
+  type NodeHash = ByteString
+  type NodeEncoded = Array[Byte]
+}
