+++ src/main/scala/io/iohk/ethereum/mpt/MptVisitors/MptConstructionVisitor.scala
@@ -1,7 +1,7 @@
 package io.iohk.ethereum.mpt.MptVisitors
 
 import io.iohk.ethereum.db.storage.MptStorage
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.mpt.BranchNode
 import io.iohk.ethereum.mpt.ExtensionNode
 import io.iohk.ethereum.mpt.HashNode
