+++ src/test/scala/io/iohk/ethereum/db/storage/StateStorageSpec.scala
@@ -15,8 +15,8 @@ import io.iohk.ethereum.db.cache.Cache
 import io.iohk.ethereum.db.cache.LruCache
 import io.iohk.ethereum.db.cache.MapCache
 import io.iohk.ethereum.db.dataSource.EphemDataSource
-import io.iohk.ethereum.db.storage.NodeStorage.NodeEncoded
-import io.iohk.ethereum.db.storage.NodeStorage.NodeHash
+import io.iohk.ethereum.db.storage.StorageTypes.NodeEncoded
+import io.iohk.ethereum.db.storage.StorageTypes.NodeHash
 import io.iohk.ethereum.db.storage.pruning.ArchivePruning
 import io.iohk.ethereum.db.storage.pruning.BasicPruning
 import io.iohk.ethereum.db.storage.pruning.InMemoryPruning
